+++ internal/api/helpers.go
@@ -5,6 +5,6 @@ import (
 )
 
 func FeesToFeeModel(minMiningFee float64) *feemodel.SatoshisPerKilobyte {
-	satoshisPerKB := int(minMiningFee * 1e8)
-	return &feemodel.SatoshisPerKilobyte{Satoshis: uint64(satoshisPerKB)}
+	satoshisPerKB := uint64(minMiningFee * 1e8)
+	return &feemodel.SatoshisPerKilobyte{Satoshis: satoshisPerKB}
 }
